+++ backend/internal/api/auth/get_profile.go
@@ -3,7 +3,7 @@ package auth
 import (
 	"context"
 	"database/sql"
-	"musicclubbot/backend/internal/helpers"
+	"musicclubbot/backend/pkg/helpers"
 	"musicclubbot/backend/proto"
 	"time"
