+++ backend/internal/api/event/get.go
@@ -3,7 +3,7 @@ package event
 import (
 	"context"
 	"database/sql"
-	"musicclubbot/backend/internal/helpers"
+	"musicclubbot/backend/pkg/helpers"
 	"musicclubbot/backend/proto"
 
 	"google.golang.org/grpc/codes"
