+++ backend/internal/api/song/join.go
@@ -3,7 +3,7 @@ package song
 import (
 	"context"
 	"database/sql"
-	"musicclubbot/backend/internal/helpers"
+	"musicclubbot/backend/pkg/helpers"
 	"musicclubbot/backend/proto"
 
 	"google.golang.org/grpc/codes"
