+++ backend/internal/api/song/list.go
@@ -3,7 +3,7 @@ package song
 import (
 	"context"
 	"database/sql"
-	"musicclubbot/backend/internal/helpers"
+	"musicclubbot/backend/pkg/helpers"
 	"musicclubbot/backend/proto"
 	"strconv"
