+++ backend/pkg/helpers/thumbnail.go
