+++ go.work
@@ -1,0 +1,5 @@
+go 1.25
+
+use (
+	"./backend"
+)
